src/test/java/org/algorithms/coding_patterns/educative/breadth_first_search/LevelAverageTest.java
package org.algorithms.coding_patterns.educative.breadth_first_search;

import org.junit.jupiter.api.Test;

import java.util.List;

import static org.junit.jupiter.api.Assertions.*;

class LevelAverageTest {

    @Test
    void findLevelAverages() {
        TreeNode root = new TreeNode(12);
        root.left = new TreeNode(7);
        root.right = new TreeNode(1);
        root.left.left = new TreeNode(9);
        root.left.right = new TreeNode(2);
        root.right.left = new TreeNode(10);
        root.right.right = new TreeNode(5);
        List<Double> result = LevelAverage.findLevelAverages(root);
        assertEquals(3, result.size());
        assertEquals(12.0, result.get(0));
        assertEquals(4.0, result.get(1));
        assertEquals(6.5, result.get(2));
    }
}